Add files via upload
This program consists of two classes - UserProfile and DatingPlatform. A user profile is created with attributes like name, age, location, and interests. The dating platform class manages a list of user profiles, it can add a new user and search users based on their interests.

In the sample usage, three users are added with different interests. Then, a search is made for users with "Sports" as an interest. All users that have "Sports" listed as an interest are printed.